requirements.txt
discord.pyyt-dlpPyNaClrequestsFlask